requirements.txt
# Streamlit Framework
streamlit==1.35.0

# AI/ML Libraries
google-generativeai>=0.3.0
anthropic>=0.18.0
openai>=1.0.0

# Web Scraping
selenium>=4.0.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
plotly>=5.18.0
kaleido>=0.2.1
openpyxl>=3.1.0
python-docx>=1.0.0

# Storage
PyGithub>=2.1.0

# Configuration
python-dotenv>=1.0.0
